fix(ui): added Loading view when page is doing full load rather than show login view
Prior to this change, during a full-page load, the user would briefly be presented with a login view
which would disappear once the loading is complete. This change adds a loading spinner until either
the page finishes loading or the site determines it's not actually logged in, and shows the login
view then.

re #37